pyproject.toml
[tool.poetry]
name = "medium-collector"
version = "0.0.0"
description = ""
authors = ["Antonio Feregrino <antonio.feregrino@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "^4.8"
imapclient = "^2.1"
python-decouple = "^3.3"
click = "^7.0"
lxml = "^4.5"
kaggle = "^1.5"

[tool.poetry.dev-dependencies]
rope = "^0.16.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
isort = "^4.3"
pylint = "^2.4"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
